comentarios

+++ atividade06.py
@@ -1,7 +1,9 @@
 # Digite um numero interio e veja sua tabuada
 
-numero = float(input("Digite um numero \n"))
+#recebendo o valor inteiro para calcular 
+numero = int(input("Digite um numero \n"))
 
+#calculando sua tabuada de 1 a 9
 t1 = numero * 1
 t2 = numero * 2
 t3 = numero * 3
@@ -11,7 +13,7 @@ t6 = numero * 6
 t7 = numero * 7
 t8 = numero * 8
 t9 = numero * 9
-
+#exibir valores
 print(t1)
 print(t2)
 print(t3)
